+++ lib/tiled/model/tiled_item_tile_set.dart
@@ -6,6 +6,7 @@ class TiledItemTileSet {
   final TileModelSprite? sprite;
   final List<CollisionArea>? collisions;
   final String? type;
+  final String? tileClass;
   final Map<String, dynamic>? properties;
   final double angle;
   final bool isFlipVertical;
@@ -16,6 +17,7 @@ class TiledItemTileSet {
     this.collisions,
     this.animation,
     this.type,
+    this.tileClass,
     this.properties,
     this.angle = 0,
     this.isFlipVertical = false,